SI-7707 Case apply transform private access
Case.apply transform is allowed from companion,
even if constructor has restricted access.

Default args are also transformed, from the
apply defaults to the ctor defaults. Usually
those expressions are identical.

Unused check is moved to refcheck, since it
is a ref check, and since the check must be
done after the case.apply transform.
