+++ src/compiler/scala/tools/nsc/ast/Trees.scala
@@ -136,7 +136,7 @@ trait Trees extends scala.reflect.internal.Trees { self: Global =>
   }
 
   class Transformer extends super.Transformer {
-    def transformUnit(unit: CompilationUnit) {
+    def transformUnit(unit: CompilationUnit): Unit = {
       try unit.body = transform(unit.body)
       catch {
         case ex: Exception =>
