+++ src/compiler/scala/tools/nsc/typechecker/Analyzer.scala
@@ -102,10 +102,7 @@ trait Analyzer extends AnyRef
           unit.body = typer.typed(unit.body)
           if (global.settings.Yrangepos && !global.reporter.hasErrors) global.validatePositions(unit.body)
           for (workItem <- unit.toCheck) workItem()
-          if (settings.warnUnusedImport)
-            warnUnusedImports(unit)
-          if (settings.warnUnused)
-            typer checkUnused unit
+          if (settings.warnUnusedImport) warnUnusedImports(unit)
         }
         finally {
           unit.toCheck.clear()
