+++ src/compiler/scala/tools/nsc/typechecker/TypeDiagnostics.scala
@@ -441,119 +441,6 @@ trait TypeDiagnostics {
     def permanentlyHiddenWarning(pos: Position, hidden: Name, defn: Symbol) =
       context.warning(pos, "imported `%s' is permanently hidden by definition of %s".format(hidden, defn.fullLocationString))
 
-    object checkUnused {
-      val ignoreNames: Set[TermName] = Set(TermName("readResolve"), TermName("readObject"), TermName("writeObject"), TermName("writeReplace"))
-
-      class UnusedPrivates extends Traverser {
-        val defnTrees = ListBuffer[MemberDef]()
-        val targets   = mutable.Set[Symbol]()
-        val setVars   = mutable.Set[Symbol]()
-        val treeTypes = mutable.Set[Type]()
-
-        def defnSymbols = defnTrees.toList map (_.symbol)
-        def localVars   = defnSymbols filter (t => t.isLocalToBlock && t.isVar)
-
-        def qualifiesTerm(sym: Symbol) = (
-             (sym.isModule || sym.isMethod || sym.isPrivateLocal || sym.isLocalToBlock)
-          && !nme.isLocalName(sym.name)
-          && !sym.isParameter
-          && !sym.isParamAccessor       // could improve this, but it's a pain
-          && !sym.isEarlyInitialized    // lots of false positives in the way these are encoded
-          && !(sym.isGetter && sym.accessed.isEarlyInitialized)
-        )
-        def qualifiesType(sym: Symbol) = !sym.isDefinedInPackage
-        def qualifies(sym: Symbol) = (
-             (sym ne null)
-          && (sym.isTerm && qualifiesTerm(sym) || sym.isType && qualifiesType(sym))
-        )
-
-        override def traverse(t: Tree): Unit = {
-          t match {
-            case t: MemberDef if qualifies(t.symbol)   => defnTrees += t
-            case t: RefTree if t.symbol ne null        => targets += t.symbol
-            case Assign(lhs, _) if lhs.symbol != null  => setVars += lhs.symbol
-            case _                                     =>
-          }
-          // Only record type references which don't originate within the
-          // definition of the class being referenced.
-          if (t.tpe ne null) {
-            for (tp <- t.tpe ; if !treeTypes(tp) && !currentOwner.ownerChain.contains(tp.typeSymbol)) {
-              tp match {
-                case NoType | NoPrefix    =>
-                case NullaryMethodType(_) =>
-                case MethodType(_, _)     =>
-                case _                    =>
-                  log(s"$tp referenced from $currentOwner")
-                  treeTypes += tp
-              }
-            }
-            // e.g. val a = new Foo ; new a.Bar ; don't let a be reported as unused.
-            t.tpe.prefix foreach {
-              case SingleType(_, sym) => targets += sym
-              case _                 =>
-            }
-          }
-          super.traverse(t)
-        }
-        def isUnusedType(m: Symbol): Boolean = (
-              m.isType
-          && !m.isTypeParameterOrSkolem // would be nice to improve this
-          && (m.isPrivate || m.isLocalToBlock)
-          && !(treeTypes.exists(tp => tp exists (t => t.typeSymbolDirect == m)))
-        )
-        def isUnusedTerm(m: Symbol): Boolean = (
-             (m.isTerm)
-          && (m.isPrivate || m.isLocalToBlock)
-          && !targets(m)
-          && !(m.name == nme.WILDCARD)              // e.g. val _ = foo
-          && !ignoreNames(m.name.toTermName)        // serialization methods
-          && !isConstantType(m.info.resultType)     // subject to constant inlining
-          && !treeTypes.exists(_ contains m)        // e.g. val a = new Foo ; new a.Bar
-        )
-        def unusedTypes = defnTrees.toList filter (t => isUnusedType(t.symbol))
-        def unusedTerms = defnTrees.toList filter (v => isUnusedTerm(v.symbol))
-        // local vars which are never set, except those already returned in unused
-        def unsetVars = localVars filter (v => !setVars(v) && !isUnusedTerm(v))
-      }
-
-      def apply(unit: CompilationUnit) = {
-        val p = new UnusedPrivates
-        p traverse unit.body
-        val unused = p.unusedTerms
-        unused foreach { defn: DefTree =>
-          val sym             = defn.symbol
-          val pos = (
-            if (defn.pos.isDefined) defn.pos
-            else if (sym.pos.isDefined) sym.pos
-            else sym match {
-              case sym: TermSymbol => sym.referenced.pos
-              case _               => NoPosition
-            }
-          )
-          val why = if (sym.isPrivate) "private" else "local"
-          val what = (
-            if (sym.isDefaultGetter) "default argument"
-            else if (sym.isConstructor) "constructor"
-            else if (sym.isVar || sym.isGetter && (sym.accessed.isVar || (sym.owner.isTrait && !sym.hasFlag(STABLE)))) "var"
-            else if (sym.isVal || sym.isGetter && (sym.accessed.isVal || (sym.owner.isTrait && sym.hasFlag(STABLE))) || sym.isLazy) "val"
-            else if (sym.isSetter) "setter"
-            else if (sym.isMethod) "method"
-            else if (sym.isModule) "object"
-            else "term"
-          )
-          reporter.warning(pos, s"$why $what in ${sym.owner} is never used")
-        }
-        p.unsetVars foreach { v =>
-          reporter.warning(v.pos, s"local var ${v.name} in ${v.owner} is never set - it could be a val")
-        }
-        p.unusedTypes foreach { t =>
-          val sym = t.symbol
-          val why = if (sym.isPrivate) "private" else "local"
-          reporter.warning(t.pos, s"$why ${sym.fullLocationString} is never used")
-        }
-      }
-    }
-
     object checkDead {
       private val exprStack: mutable.Stack[Symbol] = mutable.Stack(NoSymbol)
       // The method being applied to `tree` when `apply` is called.
