+++ test/files/pos/t7707.flags
@@ -1,0 +1,1 @@
+-Xfatal-warnings -Ywarn-unused -Xprint:typer,refcheck
